Moved DetailRow/DetailHeader to Own Component
