+++ src/components/DetailRow/index.css
