+++ src/components/DetailRow/index.ts
@@ -1,0 +1,2 @@
+export { Row as DetailRow, Props as DetailRowProps } from './Row';
+export { Header as DetailHeader } from './Header';
