--- src/components/SongList/index.ts
@@ -1,2 +1,0 @@
-export { default as SongList } from './SongList';
-export { Row as DetailRow, Header as DetailHeader } from './Detail';
